hpn #025 rename folders for some tests
