+++ sources/tests/Eshva.DockerCompose.Tests.Unit/Commands/DownProject/GivenDownProjectCommandWhenExecuteCalled.cs
@@ -11,7 +11,7 @@ using Xunit;
 #endregion
 
 
-namespace Eshva.DockerCompose.Tests.Unit.Commands.Down
+namespace Eshva.DockerCompose.Tests.Unit.Commands.DownProject
 {
     public sealed class GivenDownProjectCommandWhenExecuteCalled
     {
