+++ sources/tests/Eshva.DockerCompose.Tests.Unit/Commands/UpProject/GivenUpProjectCommandBuilderWhenBuildingCommandAndSettingOptions.cs
@@ -9,7 +9,7 @@ using Xunit;
 #endregion
 
 
-namespace Eshva.DockerCompose.Tests.Unit.Commands.Up
+namespace Eshva.DockerCompose.Tests.Unit.Commands.UpProject
 {
     public sealed class GivenUpProjectCommandBuilderWhenBuildingCommandAndSettingOptions
     {
